Move endpoints to coordination controller
